Refactor import statements and remove commented-out code in check_missing_forms
